add overarching width div to blog inherited blades
In the original 'blog' template, the width of various blades/pages (about, blog, index, contact) was set by the master blade. Since the new docs site brings in the blog, and the docs site doesn't include the sizing in the master blade, we adjust each inherited blade from blot to include such a width sizing div.